backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
numpy>=1.24.0
python-multipart>=0.0.6
pillow>=10.0.0
requests>=2.31.0
sarvamai>=0.0.1
